# service/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GridWise API
  version: 0.1.0
paths:
  /api/v1/drivers/:
    get:
      tags:
      - drivers
      summary: Get All Drivers
      description: Get all drivers with optional filters
      operationId: get_all_drivers_api_v1_drivers__get
      parameters:
      - name: team_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Team Name
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(active|inactive|reserve)$
          - type: 'null'
          title: Status
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriverResponse'
                title: Response Get All Drivers Api V1 Drivers  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - drivers
      summary: Create Driver
      description: Create a new driver
      operationId: create_driver_api_v1_drivers__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/drivers/{driver_id}:
    get:
      tags:
      - drivers
      summary: Get Driver
      description: Get a specific driver by ID
      operationId: get_driver_api_v1_drivers__driver_id__get
      parameters:
      - name: driver_id
        in: path
        required: true
        schema:
          type: string
          title: Driver Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - drivers
      summary: Update Driver
      description: Update an existing driver
      operationId: update_driver_api_v1_drivers__driver_id__put
      parameters:
      - name: driver_id
        in: path
        required: true
        schema:
          type: string
          title: Driver Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - drivers
      summary: Delete Driver
      description: Delete a driver (soft delete by setting status=inactive)
      operationId: delete_driver_api_v1_drivers__driver_id__delete
      parameters:
      - name: driver_id
        in: path
        required: true
        schema:
          type: string
          title: Driver Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/drivers/team/{team_name}:
    get:
      tags:
      - drivers
      summary: Get Drivers By Team
      description: Get all drivers for a specific team
      operationId: get_drivers_by_team_api_v1_drivers_team__team_name__get
      parameters:
      - name: team_name
        in: path
        required: true
        schema:
          type: string
          title: Team Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriverResponse'
                title: Response Get Drivers By Team Api V1 Drivers Team  Team Name  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/constructors/:
    get:
      tags:
      - constructors
      summary: Get All Constructors
      description: Get all constructors with optional filters
      operationId: get_all_constructors_api_v1_constructors__get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(active|inactive)$
          - type: 'null'
          title: Status
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConstructorResponse'
                title: Response Get All Constructors Api V1 Constructors  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - constructors
      summary: Create Constructor
      description: Create a new constructor
      operationId: create_constructor_api_v1_constructors__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConstructorCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConstructorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/constructors/{constructor_id}:
    get:
      tags:
      - constructors
      summary: Get Constructor
      description: Get a specific constructor by ID
      operationId: get_constructor_api_v1_constructors__constructor_id__get
      parameters:
      - name: constructor_id
        in: path
        required: true
        schema:
          type: string
          title: Constructor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConstructorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - constructors
      summary: Update Constructor
      description: Update an existing constructor
      operationId: update_constructor_api_v1_constructors__constructor_id__put
      parameters:
      - name: constructor_id
        in: path
        required: true
        schema:
          type: string
          title: Constructor Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConstructorUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConstructorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - constructors
      summary: Delete Constructor
      description: Delete a constructor (soft delete by setting status=inactive)
      operationId: delete_constructor_api_v1_constructors__constructor_id__delete
      parameters:
      - name: constructor_id
        in: path
        required: true
        schema:
          type: string
          title: Constructor Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ConstructorCreate:
      properties:
        name:
          type: string
          title: Name
        full_name:
          type: string
          title: Full Name
        nationality:
          type: string
          title: Nationality
        price:
          type: number
          title: Price
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - name
      - full_name
      - nationality
      - price
      title: ConstructorCreate
    ConstructorResponse:
      properties:
        name:
          type: string
          title: Name
        full_name:
          type: string
          title: Full Name
        nationality:
          type: string
          title: Nationality
        price:
          type: number
          title: Price
        status:
          type: string
          title: Status
          default: active
        constructor_id:
          type: string
          title: Constructor Id
        price_history:
          items:
            $ref: '#/components/schemas/PriceHistory'
          type: array
          title: Price History
          default: []
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - name
      - full_name
      - nationality
      - price
      - constructor_id
      - created_at
      - updated_at
      title: ConstructorResponse
    ConstructorUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        nationality:
          anyOf:
          - type: string
          - type: 'null'
          title: Nationality
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: ConstructorUpdate
    DriverCreate:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        team_name:
          type: string
          title: Team Name
        nationality:
          type: string
          title: Nationality
        driver_number:
          type: integer
          title: Driver Number
        price:
          type: number
          title: Price
        status:
          type: string
          title: Status
          default: active
      type: object
      required:
      - first_name
      - last_name
      - team_name
      - nationality
      - driver_number
      - price
      title: DriverCreate
    DriverResponse:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        team_name:
          type: string
          title: Team Name
        nationality:
          type: string
          title: Nationality
        driver_number:
          type: integer
          title: Driver Number
        price:
          type: number
          title: Price
        status:
          type: string
          title: Status
          default: active
        driver_id:
          type: string
          title: Driver Id
        price_history:
          items:
            $ref: '#/components/schemas/PriceHistory'
          type: array
          title: Price History
          default: []
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - first_name
      - last_name
      - team_name
      - nationality
      - driver_number
      - price
      - driver_id
      - created_at
      - updated_at
      title: DriverResponse
    DriverUpdate:
      properties:
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        team_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Team Name
        nationality:
          anyOf:
          - type: string
          - type: 'null'
          title: Nationality
        driver_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Driver Number
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: DriverUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PriceHistory:
      properties:
        price:
          type: number
          title: Price
        changed_at:
          type: string
          format: date-time
          title: Changed At
        changed_by:
          type: string
          title: Changed By
          default: system
      type: object
      required:
      - price
      - changed_at
      title: PriceHistory
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...